from Dev to master

+++ src/python/module/nifty/graph/rag/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import
+from .._graph import *
 from ._rag import *
 from .. import Configuration
 
@@ -11,7 +12,7 @@ for key in _rag.__dict__.keys():
 
 
 def gridRag(labels, numberOfThreads=-1, serialization = None):
-    labels = numpy.require(labels)
+    labels = numpy.require(labels ,dtype='uint32')
 
 
     if numpy.squeeze(labels).ndim == 2:
